fix: testing all of methodes user in package repository
